Declare 'gridCell' with let instead of const
Previously gridCell was defined with const, however this meant its value
could not be reassigned, further meaning that gridCell had to declared
in the for loop during every iterations(16 times).

Now, declaring gridCell with 'let' allows it to be assigned a new div
during each iteration outside the for loop as its value can change
allowing its use elsewhere in the program.